Remove the option, update readme, edit file order
